Add substance implemented

+++ src/main/java/ohtu/radioaine/database/DAO.java
@@ -7,7 +7,7 @@ package ohtu.radioaine.database;
  */
 public interface DAO<T> {
     public T create(T instance);
-    public T read(long id);
+    public T read(int id);
     public void delete(T instance);
     public T update(T instance);
 }
